refactor: modularize stream-v2 (#3594)
