+++ open-sse/utils/stream/openaiChunks.ts
@@ -1,0 +1,7 @@
+import { JsonRecord } from "./types.ts";
+
+export function getOpenAIIntermediateChunks(value: unknown): unknown[] {
+  if (!value || typeof value !== "object") return [];
+  const candidate = (value as JsonRecord)._openaiIntermediate;
+  return Array.isArray(candidate) ? candidate : [];
+}
